+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added the pkg.pr.new workflow
+- The minLength and maxLength input properties
+
 ## [2.12.0] - 2024-10-07
 
 ### Added